+++ wasm/src/dbg/handler.rs
@@ -26,7 +26,8 @@ use clvm_tools_rs::compiler::sexp::{decode_string, parse_sexp, SExp};
 use clvm_tools_rs::compiler::srcloc::Srcloc;
 
 use crate::dbg::compopts::DbgCompilerOpts;
-use crate::dbg::obj::{parse_srcloc, read_program_data, RunningDebugger, StoredScope, TargetDepth};
+use crate::dbg::obj::{read_program_data, RunningDebugger, TargetDepth};
+use crate::dbg::source::{StoredScope, parse_srcloc};
 use crate::dbg::types::MessageHandler;
 use crate::interfaces::{IFileReader, ILogWriter};
 use crate::lsp::types::ConfigJson;
