+++ wasm/src/dbg/mod.rs
@@ -2,4 +2,5 @@ pub mod compopts;
 pub mod handler;
 pub mod obj;
 pub mod server;
+pub mod source;
 pub mod types;
